Cast eval features to float32 before matmul/cosine/CKA
Under AMP (enable_amp=True, amp_dtype="bfloat16") the Qwen ViT output is
bfloat16 while patch_proj's output is float32, so downstream tensor ops
mixed the two dtypes and raised:

  RuntimeError: expected scalar type Float but found BFloat16
  ... at sim = f3n @ f2n.T (retrieval) / X.T @ Y (CKA)

Cast both feature pairs to float32 right after extraction so all
downstream math (cosine_similarity, retrieval matmul, CKA) runs in a
single dtype.

Also swapped point_collate_fn invocation in eval_alignment_full.py for
a tighter `_coerce_sample_for_model` helper that just promotes scalar
floats/ints (notably `grid_size: 0.01`) to 1-element tensors — avoids
relying on Pointcept's collate machinery for the single-sample eval path.

+++ training/tools/eval_alignment.py
@@ -45,7 +45,7 @@ matplotlib.use("Agg")
 import matplotlib.pyplot as plt
 
 from pointcept.engines.defaults import default_config_parser
-from pointcept.datasets import build_dataset, point_collate_fn
+from pointcept.datasets import build_dataset
 from pointcept.models import build_model
 from pointcept.models.utils import offset2batch, bincount2offset
 from pointcept.models.utils.structure import Point
@@ -170,6 +170,11 @@ def _eval_one_sample(model, data_dict, device, generator):
         feature_index_unique = torch.unique(feature_index)
         f2 = feature2d[feature_index_unique]
         f3 = feature3d_pixel[feature_index_unique]
+
+        # Cast to float32 — AMP-bfloat16 Qwen output vs float32 patch_proj
+        # output otherwise fails the cosine matmul with a dtype mismatch.
+        f2 = f2.float()
+        f3 = f3.float()
 
         if getattr(model, "enc2d_cos_shift", False):
             f2 = f2 - f2.mean(dim=-1, keepdim=True)
